requirements.txt
keyboard==0.13.5
matplotlib==3.8.3
mlxtend==0.23.1
numpy==1.26.4
pandas==2.2.1
PyAutoGUI==0.9.54
torch==2.2.1
torchinfo==1.8.0
torchmetrics==1.3.1
torchvision==0.17.1
tqdm==4.66.2
python-box==7.1.1